+++ src/test/kotlin/tornadofx/testapps/PojoTreeTableColumns.kt
@@ -10,7 +10,7 @@ class PojoTreeTableColumns : View("Pojo Tree Table Columns") {
 
     val people = listOf(
             JavaPerson("Mary Hanes", "IT Administration", "mary.hanes@contoso.com", "mary2.hanes@contoso.com", listOf(
-                    JavaPerson("Jacob Mays", "IT Help Desk", "jacob.mays@contoso.com", "jacob2.mays@contoso.com" ),
+                    JavaPerson("Jacob Mays", "IT Help Desk", "jacob.mays@contoso.com", "jacob2.mays@contoso.com"),
                     JavaPerson("John Ramsy", "IT Help Desk", "john.ramsy@contoso.com", "john2.ramsy@contoso.com"))),
             JavaPerson("Erin James", "Human Resources", "erin.james@contoso.com", "erin2.james@contoso.com", listOf(
                     JavaPerson("Erlick Foyes", "Customer Service", "erlick.foyes@contoso.com", "erlick2.foyes@contoso.com"),
@@ -19,11 +19,13 @@ class PojoTreeTableColumns : View("Pojo Tree Table Columns") {
     ).observable()
 
     // Create the root item that holds all top level employees
-    val rootItem = TreeItem( JavaPerson().apply{ name = "Employees by Manager"; employees = people  })
+    val rootItem = TreeItem(JavaPerson().apply { name = "Employees by Manager"; employees = people })
 
-    override val root = treetableview(rootItem){
-        column<JavaPerson,String>("Name", "name" ) // non type safe
-        column("Department", JavaPerson::getDepartment )
+    override val root = treetableview(rootItem) {
+        prefWidth = 800.0
+
+        column<JavaPerson, String>("Name", "name").contentWidth(50) // non type safe
+        column("Department", JavaPerson::getDepartment).remainingWidth()
         nestedColumn("Email addresses") {
             column("Primary Email", JavaPerson::getPrimaryEmail)
             column("Secondary Email", JavaPerson::getSecondaryEmail)
@@ -35,6 +37,6 @@ class PojoTreeTableColumns : View("Pojo Tree Table Columns") {
         // Expand the two first levels
         root.isExpanded = true
         root.children.forEach { it.isExpanded = true }
-        resizeColumnsToFitContent()
+        smartResize()
     }
 }